// starter-frontend/src/app/api/generated/model/format-enum.ts
/**
 * Starter API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * * `text` - text * `html` - html
 */
export enum FormatEnumDto {

    Text = 'text',

    Html = 'html'
}