// starter-frontend/src/app/api/generated/model/translate-batch-request-request.ts
/**
 * Starter API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { TranslateItemRequestDto } from './translate-item-request';


export interface TranslateBatchRequestRequestDto { 
    source: string;
    target: string;
    items: Array<TranslateItemRequestDto>;
}