// starter-frontend/src/app/api/generated/model/translate-item-request.ts
/**
 * Starter API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { FormatEnumDto } from './format-enum';


export interface TranslateItemRequestDto { 
    key: string;
    text: string;
    format?: FormatEnumDto;
}